rearrange assessment route tests
